feat: Add Header component and update layout
This commit adds a new Header component to the layout, which is imported from the containers directory. It also updates the layout to include the Header component and removes unused imports. Additionally, the global styles have been modified to include a box-sizing property and set the max-width and overflow-x properties for the body element. Finally, a new Nav component has been added to the components directory.

chore: Add Sass dependency

This commit adds the Sass dependency to the project.

Revert "chore: Add product type definition"

This commit reverts the previous commit that added the product type definition.

Initial commit from Create Next App

This is the initial commit of the project, created using Create Next App.