feat: Add API route for fetching products
This commit adds a new API route for fetching products. It includes the implementation of the GET method in the `route.ts` file, which returns a JSON response containing the `products` data from the `testData.ts` file. This will allow clients to retrieve the list of products from the server.